Refactored ThemeStore localStorage logic a bit
